Added filter list to dan; Updated filter phrasing
Dan now uses a filter list. By default it only has the rating:safe tag.
Unfortunately this does count toward the 'tag limit'. It can of course
be removed using the usual filter manipulation command. resolves #2

Updated the add filter command to specify it's adding to the server.
Updated the del filter command to specify it's reverting for the server.